refactor: move Product interface declaration to hook & change imports
